Create helper function folder and move users helper function into it
